update testing scripts

+++ terra/setup.sh
@@ -9,14 +9,18 @@ mkdir $SEQREPO_ROOT
 seqrepo --root-directory $SEQREPO_ROOT pull --update-latest
 
 # setup bcftools
-cd ~
-curl -LJO https://github.com/samtools/bcftools/releases/download/1.19/bcftools-1.19.tar.bz2
-BCFTOOLS_TAR=$(ls -1t bcftools*.tar.bz2 | head -n 1)
-tar -xvf $BCFTOOLS_TAR
-rm $BCFTOOLS_TAR
+if [ bcftools --version &> /dev/null ]; then
+    cd ~
+    curl -LJO https://github.com/samtools/bcftools/releases/download/1.19/bcftools-1.19.tar.bz2
+    BCFTOOLS_TAR=$(ls -1t bcftools*.tar.bz2 | head -n 1)
+    tar -xvf $BCFTOOLS_TAR
+    rm $BCFTOOLS_TAR
 
-mv ~/bcftools*/ $BCFTOOLS_DIR
-cd $BCFTOOLS_DIR
-./configure prefix=$HOME
-make
-make install
+    mv ~/bcftools*/ $BCFTOOLS_DIR
+    cd $BCFTOOLS_DIR
+    ./configure prefix=$HOME
+    make
+    make install
+else
+    echo "bcftools already installed, not installing"
+fi
